requirements.txt
fastapi
uvicorn
groq
pytesseract
Pillow
edge-tts
python-dotenv
python-multipart
